% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ensemble.machine.learning.thin.plate.splines.V73.R
\name{machisplin.tiles.id}
\alias{machisplin.tiles.id}
\title{Show diagram of sub-sampling grid}
\usage{
machisplin.tiles.id(rast.in)
}
\arguments{
\item{rast.in}{output from machisplin.tiles.create tool}
}
\value{
This function displays the order of analysis - rasters input into machisplin.tiles.merge rast.in parameter MUST match this order.
}
\description{
Show diagram of sub-sampling grid
}
\examples{
library(MACHISPLIN)

# Import a csv 
Mydata	<-read.delim("sampling.csv", sep=",", h=T)

#load rasters to use as high resolution co-variates for downscaling
ALT = rast("SRTM30m.tif")
SLOPE = rast("ln_slope.tif")
ASPECT = rast("aspect.tif")
TWI = rast("TWI.tif")

# function input: raster stack of covarites
raster_stack<-c(ALT,SLOPE,TWI,ASPECT)

#sub-divide landscape into smaller units to facilite downscaling
tile<-machisplin.tiles.create(rast.in= raster_stack, int.values=Mydata,out.ncol=2, out.nrow=2, feather.d=50)

#show grid
machisplin.tiles.id(tile)
}
