% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ensemble.machine.learning.thin.plate.splines.V71.R
\name{machisplin.write.loadings}
\alias{machisplin.write.loadings}
\title{Write MACHISPLIN model loadings}
\usage{
machisplin.write.loadings(mltps.in, out.names = NULL)
}
\arguments{
\item{out.names}{a vector corresponding to the output raster names.   If 'null' it will write intial input names.}

\item{mttps.in}{an output from the 'machisplin.mltps' function}
}
\value{
This function outputs performance values, the algorithm(s) used, and rasters for use in GIS output from the 'machisplin.mltps' function
}
\description{
Write MACHISPLIN model loadings
}
\examples{
library(MACHISPLIN)

# Import a csv 
Mydata	<-read.delim("sampling.csv", sep=",", h=T)

#load rasters to use as high resolution co-variates for downscaling
ALT = rast("SRTM30m.tif")
SLOPE = rast("ln_slope.tif")
ASPECT = rast("aspect.tif")
TWI = rast("TWI.tif")

# function input: raster stack of covarites
raster_stack<-c(ALT,SLOPE,TWI,ASPECT)

#run an ensemble machine learning thin plate spline 
interp.rast<-machisplin.mltps(int.values=Mydata, covar.ras=raster_stack)

machisplin.write.loadings(mltps.in=interp.rast)
}
