% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ensemble.machine.learning.thin.plate.splines.V19.R
\name{machisplin.mltps}
\alias{machisplin.mltps}
\title{Machine Learning Ensemble & Thin-Plate-Spline Interpolation}
\usage{
machisplin.mltps(int.values, covar.ras, n.cores = 1)
}
\arguments{
\item{imported}{table for analysis}
}
\value{
This tool remove NAs and converts all input data to numeric values for analysis in Humboldt.
}
\description{
Machine Learning Ensemble & Thin-Plate-Spline Interpolation
}
\examples{
library(machisplin)

library(raster)
library(gstat)
library(maptools)
library(sp)
library(rgdal)
library(spdep)
library(nlme)
library(MASS)
library(fields)
library(snowfall)
library(dismo)
library(randomForest)
library(earth)
library(kernlab)
library(mgcv)

# Import a csv as shapefile:
Mydata<-read.delim("sampling.csv", sep=",", h=T)

#load rasters
ALT = raster("SRTM30m.tif")
CURV_V = raster("cs_curv.tif")
CURV_H = raster("long_cur.tif")
SLOPE = raster("ln_slope.tif")
REL_ALT= raster("relative_elevation500m.tif")
TWI = raster("TWI.tif")

# function input: raster brick of covarites
raster_stack<-stack(ALT,REL_ALT,SLOPE,TWI,CURV_V,CURV_H)

#run an ensemble machine learning thin plate spline 
interp.rast<-machisplin.mltps(int.values=Mydata, covar.ras=raster_stack, n.cores=2)

}
