% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ensemble.machine.learning.thin.plate.splines.V20.R
\name{machisplin.write.geotiff}
\alias{machisplin.write.geotiff}
\title{Write MACHISPLIN geotiff}
\usage{
machisplin.write.geotiff(mltps.in, out.names = NULL)
}
\arguments{
\item{out.names}{a vector corresponding to the output raster names.   If 'null' it will write intial input names.}

\item{mttps.in}{an output from the 'machisplin.mltps' function}
}
\value{
This function outputs performance values, the algorithm(s) used, and rasters for use in GIS output from the 'machisplin.mltps' function
}
\description{
Write MACHISPLIN geotiff
}
\examples{
library(MACHISPLIN)

# Import a csv 
Mydata	<-read.delim("sampling.csv", sep=",", h=T)

#load rasters				 
ALT = raster("SRTM30m.tif")
CURV_V = raster("cs_curv.tif")
CURV_H = raster("long_cur.tif")
SLOPE = raster("ln_slope.tif")
REL_ALT= raster("relative_elevation500m.tif")
TWI = raster("TWI.tif")

# function input: raster stack of covarites
raster_stack<-stack(ALT,REL_ALT,SLOPE,TWI,CURV_V,CURV_H)

#run an ensemble machine learning thin plate spline 
interp.rast<-machisplin.mltps(int.values=Mydata, covar.ras=raster_stack, n.cores=2)

machisplin.write.geotiff(mltps.in=interp.rast)
}
